Add figure images
These are sized at 1440px on the longest dimension. Many will require
some masking on the borders to make them clean and consistent. The
specifics of that masking, such as the color to use on borders, will be
dependent on the design and layout.

This set includes three FPO images which will need to be replaced.