Add figure markup and captions to texts
Markup formatting follows structure set for the Guide section of the
Ancient Terracottas catalogue. Including, the use of src='<%=
image_path("figures/fig.jpg") %>'

This commit also uploads one previously missing figure.

Waiting on an answer on a question to Nola regarding the formatting of
the captions. In the docs transmitted to us, there were no italics
indicated for titles, and no periods following the last credit. Not
sure if that’s correct or not.