Convert span tag formatting and add yml files for pop-ups
Throughout text, added markup for three different kids of pop-ups:
Location, Definition, and Pic. Using the following as examples:

<span class="popup location" id="loc_175">Aix-en-Provence</span>
<span class="popup definition" id="def_nymphaea">*nymphaea*</span>
<span class="popup pic" id="pic_48">Church of Saints Lot and
Procopius</span>

Also added YML files for definitions and for pics. Though I’m getting a
validation issue with the definitions.yml at this point. And a
locations data file is yet to come.

Also to come, there remain about a dozen locations references in the
text marked as {{Name}}, that need location id numbers assigned, and
need to be converted in the text to the proper span tag format.