Change catalogue.yml to a simple array
Middleman doesn't like the 1: ..., 2: ... schema we had used previously
for the data/catalogue.yml file. Going forward we should use a simple
array for these kinds of files. That means that array index and
catalogue number will be different, so array index should not be relied
on to determine this. Just use the `cat` property of each array element
instead.