Set up basic TOC
The logic in contents.html.erb is able to parse the book structure
adequately for now, understanding the difference between single and
multiple layouts, etc. Eventually this should probably be moved into
some kind of helper method.

In addition, the `data` hash of sitemap resources no longer contains
`layout` info - it looks like this is stripped out by MM. So intro
sections should have an explicit flag that defines them as such so that
the TOC knows how to treat them.