Add basic responsive nav + ionicons
This begins the build-out of the app UI. Ionicons are back (they are
unobtrusive and minimal) as is the basic fixed navbar element.

The goal for this project is to keep markup as clean as possible. So for
things like ionicons, where each icon contains its own class, these
classes should be included via sass @extends, instead of placing
non-semantic classes directly into the HTML. Same goes with utility
classes like .hide-on-mobile, etc.