Fix cutoff text
On a mid-size laptop screen, the large media query was being triggered
but content did not properly fit the display. Probably an x-large
query is going to need to be added to account for that use case;
will add this in as necessary.

Also, variable names have been shortened to try and keep things within
the 80-char limit.