Implement reusable Table of Contents nav
The TOC has been moved into a re-usable partial with basic styles; in
different contexts these styles can be partially overridden to minimize
duplication of code.

In addition to the stylesheets for this new feature, this commit
includes a bugfix. Previously the TOC template in
`source/partials/_contents_list.html.haml` linked to items by using
their `path`; this is relative, and was breaking when the point of
departure was inside another directory (like all /catalogue pages). I've
changed this to `url` and things are working. Hopefully this is still
compatible with relative links for deployment.