Enable basic navbar slideout behavior
Right now this can be switched from a left to right layout by changing
the value of the $sliding-panel-position in _nav_contents.scss.

A mobile version (full-screen slide-out) should probably be created
next. Also, there should be a button to close the menu within the menu
itself in this case.