Add page transitions
Despite all the trends towards single-page apps with routing frameworks,
etc., SmoothState remains an excellent tool for improving both
performance and user experience while embracing the spirit of
progressive enhancement. The project continues to be actively developed
and it is easy to leverage external libraries like velocity.js for
fine-grained animations.

This commit represents the most basic setup for the plugin. In the main
application template, content has been organized under a #main element
inside of the <body> tag. Everything of importance lives here
(navigation elements included).

In the main application.js file, SmoothState is initialized on this
container element. The options passed to it define the behavior at
different stages. At minimum for nice transitions you should define an
onStart and onReady object. Finally, use the onAfter hook to
re-initialize any UI elements (in this case, keyboard nav and off-canvas
sliding panel).

It's amazing how much the addition of a simple transition between pages
improves the subjective experience of using the app.