Enable search index generation
This commit adds a simple extension that generates a JSON index of the
full text of the site's contents, for use by Lunr.js in client-side
searches. In Terracottas, there is a contents.erb.json template file in
the general /views/ folder, but this is really a derivative file and
should not be treated like content in the project. Although the
extension approach is more complex it represents a better separation of
concerns in the overall project.

Much of this code was adapted from the middleman-search gem, in
particular the SearchIndexResource helper class
(https://github.com/manastech/middleman-search). But I have simplified
the code and customized it to suit the needs of this project (in
particular, there was a need to pull some content from the catalogue
data files into the search index).

The short explanation of how this works is: a SearchIndexResource class
is defined, which inherits from Middleman::Sitemap::Resource. The search
extension uses Middleman's manipulate_resource_list method to create an
instance of this class and append it to the global site resources
(giving it a path of "/search.json" by default). The SearchIndexResource
instance takes care of the rest internally (looping through resources
when they meet certain criteria, stashing the relevant text from these
resources into an array of hashes, and defining a #render method which
converts this index into a string of JSON).

Next steps are wiring up Lunr.js to use this file and building out the
search UI. In Terracottas this was done via Handlebars but I'd like to
remove that dependency and do all client-side rendering with Vue.js this
time around.

One final note: the SearchIndexResource introduces a new project
dependency, on the Sanitize gem (for cleaning up text post-render - no
need to clutter the search index with a bunch of HTML tags).