Set up links to be destination-agnostic
In Terracottas, we resorted to a hack where a site_url variable was set
based on whether the site was running locally or being built, since that
was the only way we could deal with the content being hosted one or two
directories deep (like on Github pages). This meant our links were
partially hard-coded, and when it came time to convert content to other
formats like EPUB, they had to be manually re-written.

It turns out all of this effort was unnecessary, we were being too
literal-minded. It turns out it is possible to pass a
Middleman::Sitemap::Resource object directly to the link_to helper
method, and Middleman will resolve the URL automatically. This is more
error-proof than passing the resource.url, resource.path, or
resource.destination_path (all of which fail when dealing with nested
subdirectories on Github pages).

Example:
- curr_entry = sitemap.resources.find { |p| p.data == current_page.cat }
= link_to curr_entry.title, curr_entry

(Note the lack of qualifier on the second argument)

For reference see here: https://middlemanapp.com/basics/helper_methods/