Set up search control UI
The search icon in the menu bar now does responds to click/touch events.
Right now it is responsible for adding/removing the "search-active"
class to the entire navbar (its parent element).

- Styles
A series of new style rules have been added which re-style this bar
based on the presence of this class. Elements are added/removed with
display:none, and flexbox rules control the new layout of content.
Layout appears completely responsive.

- Javascript
A new set of JS functions have been added to the ui.js file to support
this interaction. First, an improvement – the ui.js functions have been
packaged into a single function at the end of the file: uiSetup(). This
is a poor man's module.exports – as this file grows, make sure to add a
function call for every function defined above it inside the body. Then
other files only need to care about a single interface which calls all
the other things which need to be set up. It's not true encapsulation
(Webpack or Browserify would be needed for that, and there just is not
enough time) but it's better than nothing.

Otherwise there is nothing very complex happening here at the moment,
just basic click/touch event handlers to add/remove state classes. The
escape key has been bound to remove the class (if present) which
controls both sidebar active state and search active state.

Next steps are building a template to render actual search results. This
should be loaded as a partial inside of the nav element somewhere, and
ideally result display can be handled via Vue.js so that only a single
library is needed for client-side rendering. One advantage of using Vue
is that the component could wait to load search index data until the
search icon is actually clicked, so there is no delay on normal page
loads while the search index gets pulled in.