Add geojson data for locations
This JSON was derived from the Digital Atlas of the Roman Empire
(http://dare.ht.lu.se/). The “custom_name” field corresponds to how the
location is referred to in the catalogue, and should be used for
display. The “dare” id corresponds to the Atlas itself, and to the
location numbers for the image pop-ups, currently recorded in
locations.yml and tagged throughout the manuscript