Add search results panel element
This commit adds an empty .search-results panel to the nav partial. In
JS, the search-active class is added/removed from this element at the
same time that it is modified on the navbar.

Stylistically, this is a fixed-position element with a z-index of 9998
(one point below the navbar), with a height and width that fills the
viewport. By default it is translated "north" of the viewport, but when
the .search-active class is applied it slides down in a 0.1ms
transition.

Next up: build out a new partial that lives inside of this panel which
can display search results.