Client-side search with Vue and Lunr.js
This commit is a big step forward for the project. We now have working
client-side full-text search using lunr.js, as in the Terracottas
project. However, the organization of the code has been greatly improved
and the Handlebars.js dependency has been removed in favor of Vue.

This should be much easier to reason about going forward, as all
search-related logic is organized into clearly defined methods.

- Dependencies
This commit adds the vue-resource library for simplified HTTP requests
inside of vue components. This is in addition to the previously
committed dependencies on Vue.js and Lunr.js. Because this project does
not use NPM, all of these js dependencies have been committed to source.
I'm really beginning to see this as an upside, as NPM modules change so
quickly it makes it hard to be confident in the reproducibility of a
build.

- Template updates
Vue.js requires some custom attributes in templates (v-bind, @click,
etc.) similar to how Angular works. Haml seems to choke on these by
default, so all templates which are going to be used by a Vue instance
should be written in ERB. Fortunately there is no problem with mixing
partials and layouts that have been written in different languages.

One important note: the Vue instance which handles search and results
needs access to the entire .navbar component; users will enter queries
in the navbar search input, and then results will be shown in an
adjacent search-results element.

To manage this, the entire contents of the nav partial have been nested
under a #js-nav div. Also, the entire partial has been re-written in
ERB for compatibility.

- Styles
New style rules targeting .search-results, .search-results-inner, and
.search-results-list-item have been added to the .search.scss file.

- Search Logic
All search logic now lives in the search.js file, organized as a Vue
instance.

One thing lives outside of this instance in the global namespace for
now: a SearchIndex variable which is created by Lunr. This needs to
remain outside Vue because it has its own methods (like search()) which
the Search vue instance needs to be able to call.

Actual result matching is done by using Vue's list-rendering and a
computed property of searchResults. This is an array of data that
combines Lunr's index matches with an array of the actual site contents
(to get title/url to the resulting pages).

- Interaction with SmoothState
This was my main concern in using Vue for elements which will appear
across pages (instead of just on a single page), but things seem to be
working fine. The key is to manage the creation/destruction of
instances in Smoothstate's hooks. So vue instances should be created
when the page first loads (inside of $(document).ready()); they should
be destroyed (using vm.$destroy()) inside the onStart callback, and then
re-created in the onAfter callback.

To clear out the user's last search after changing pages, I have also
added a destroy hook to the vue instance which resets the query to an
empty string. This seems to provide the expected behavior without any
problems so far.