Cleanup GeoJSon file
Some wonky indentation in the geojson.js file has been cleaned up.
Additionally, this file contains GeoJSON but is actually a plain JS file
– there is variable assignment going on, etc. For now this should be
fine (the file will be required via sprockets, so the variable
containing this data will be available in the global context). This may
be something to change in the future. Renaming file to reflect what is
actually going on for now.