Add region map to section intro pages
The "intro" layout has been updated to include a map zoomed in on the
region in question. Each intro page should have a frontmatter attribute
called map_center, containing an array of lat/long values.

In terms of styles and UI changes, these maps have all had their
scrollwheel response disabled, since the idea is that the user will
scroll down the page to read the main content.

The header which appears across the bottom also needs a z-index applied
(1000 seems safe for now; 100 was too low) in order to appear above the
material that Leaflet creates upon rendering.

The intro layout has been updated to handle all of this, along with a
new partial file which contains the map div (and relevant data
attributes).