Fix Map control styles
Due to the current map layout, the leaflet-control-container element was
getting lost (attribution was not visible, for example). These style
changes fix this by forcing the container to have a position: absolute
whenever it falls inside of a parent .map-fullscreen element.

An additional rule for .map-intro pushes up this element from the bottom
by a height to match the height of the page title which appears at the
bottom.