Add overview map to cover page
This commit enables the introductory map on the cover page. Since this
map serves a different purpose than the more detailed ones later on,
there is a conditional statement in the mapSetup() function which checks
if the #map element has a .cover-map as a parent. If so, the setZoom
method is called and the map is zoomed out to level 5 (the minimum).

Since the GeoMap class is a wrapper object that stores the leaflet
instance in its `map` property, make sure to call methods on that
property directly (i.e. map.map.setZoom(), if the instance is named
'map').