Enable map markers for catalogue entries
This commit displays the location of all geoJson points with a
"catalogue" property defined as standard Leaflet markers on the map.
When clicked a basic popup containing the site custom_name and a list of
all linked entries appears.

A few changes to various files in order to get this working:

- Stylesheets
Z-index on the leaflet-control-container (increased in _global.scss
earlier) had to be taken down a bit, because the control div was making
the leaflet markers impossible to click (a bug that took a while to
figure out; I had assumed this was a JS problem but really it was due to
styling).

- GeoJSON
This data needed to be cleaned up for consistency. Going forward, the
"catalogue" property should either not exist (if there are no entries)
or it should be an array of one or more integers.