Generate a catalogue.json file programatically
This commit adds a new "catalogue" extension which generates a JSON
version of the catalogue data file for use by JS assets (primarily
Leaflet at this point).

The extension follows the structure of the search index extension
(there is an accompanying resource class in a separate file which is
used to define this resource's #render method (just a to_json call on
the catalogue data object).

In the future it may make sense to move other functionality here, even
to make this extension the "master" extension which houses all the
catalogue-specific helpers, etc. But that's for later.