Enable cat entry lookup in map popups
This commit adds underscore-min.js as a dependency (committed to
source). In order to get actual catalogue entry titles in map pop ups,
the catalogue json has been made available globally as the constant
CATALOGUE. This is less than ideal but there is no other obvious way to
access this data inside of Leaflet's callback functions.

Now, when a #map element is detected in a page, the mapSetup() function
runs and makes an ajax request to get the catalogue.json data. Inside of
this request's callback, the map is instantiated and the data is stashed
in the window.CATALOGUE global constant.

In the leaflet code, a quick call to _.find() cross-references the
catalogue number in the geojson data with the corresponding entry in the
catalogue.json data. This is used to provide the title inside of the map
popup.