Use underscore.js to access catalogue data in map
This commit completes (for now) the integration of the Leaflet map with
the catalogue entry data from the book. When linked catalogue entries
are present in a map location, some calls to underscore.js methods allow
us to easily pull out all relevant entries from the collection.

This is accomplished by the following pair of functions:

```js
var cat_ids = props.catalogue
linkedEntries = _.filter(window.CATALOGUE, function(entry) {
  return _.includes(cat_ids, entry.cat);
}
```

This uses underscore's includes() (aka contains) method to filter out a
subset of all catalogue entries based on whether or not their cat.
number is contained in the cat_ids array.

The location value of the first entry in this set is used to provide the
location text for the entire pop-up bubble.

Finally, since JS lacks a capitalize method, an underscore mixin has
been included to provide this (locally, at least).