Add an erb template to generate csv data file from catalogue.yml!
This template auto-generates the CSV file of catalogue data we offer
users for download, pulling from the catalogue data in catalogue.yml.
In previous iterations this file was hand built and so required manual
updates to keep it in synch with corrections and revisions made to the
catalogue itself.

The structure of this CSV (the data we’re collecting and the naming of
the fields) matches the standard we established with the Ancient
Terracottas catalogue.