Add template to generate json dataset for download
Like for the CSV in commit
https://github.com/gettypubs/mosaics/commit/b84e32af5e4036c0c3e3d250b20c
59ea00010bf4 this template generates the JSON file from catalogue.yml
for readers to download.

At this point, the fields here match those in the CSV, and also the
structure of the datasets made available for the Ancient Terracottas
catalogue. Further research will be done on the utility of these
formats/structures and whether there is value in having them consistent
from one catalogue to the next