Basic popup definition
This commit enables the minimum viable version of this feature: terms in
the text can be given a treatment where they receive a light underline,
and on hover a definition of the term appears.

One advantage of the current solution is that it relies 100% on CSS, no
javascript required.

- define_term(term) helper method
First, a new helper method needed to be created to make looking up the
definition of a given entry simpler in ERB templates. This method is
just an alias for a simple array#find lookup:

def define_term(term)
  data.definitions.find { |entry| entry.id == term}.definition
end

- Add definitions as data-attributes
Adding definitions to particular terms seemed like a good use-case for
HTML data attributes. This means the definitions always live with their
relevant terms and new hidden elements are not littered all over the
pages. This seems like the most accurate representation of the
underlying relationship. Add data-definitions like this:

<span class="popup popup-definition" data-definition="<%=
define_term('amphora') %>">amphora</a>

If a term has two words, make sure to separate them by a hypen (or match
whatever is written in the data.definitions.yml file).

Once this is done, CSS styles take care of the rest.