Add graphics illustrating the mosaics' scale to catalogue entries
This adds a scaled silhouette to the object details of each mosaic next
to a silhouette of a standing person at a 1px:2cm scale. Based on the
object dimensions in catalogue.yml and a person of average 170cm height.

This is primarily a demonstration intended for curatorial/editorial
feedback. It is as yet not responsive. Also, to accommodate the widest
mosaic (cat 2) the right column is somewhat awkwardly made to be 52%
width instead of 50%. There is also potentially an issue with width vs
height dimensions as recorded in the catalogue data that may
necessitate further adjustments.

There may also be some desire to have perspectival illustrations of the
mosaics on the floor rather than these more vertical views, however
that approach would require custom illustrations for each of the
mosaics to capture the correct view and dimensions. Whereas this option
is generated programatically from the catalogue data.