Add #location helper method
There is now a #location helper method which returns a location id (a
string like loc_0001, etc), though in the future this method could be
expanded to return more information (anything available in the
locations.yml file is accessible).

This data should be added as a data-attribute in the HTML files to allow
for pop-up map elements to be created.