Remove plate references
Manuscript included references to plates, but in the current version,
the plates aren’t identified as such. This in mind, and in keeping with
the precedent of past catalogues, excess plate references have been
removed altogether. Others have been edited from plate to cat. So “(see
plate 4)” for instance, becomes “(cat. 4)”