Split author listing across fold on cover
This puts the main author name on the cover, and other contributors
just below. To do this, I re-listed the author_as_it_appears field in
book.yml as an array which will allow us to establish a hierarchy of
that info and use each individual part separately. For example, were
there also a “translated by …” line, that could appear as a third item
in the array and either be concatenated with the other contributors for
display, or styled and positioned differently.