Enable basic multi-entry layout
This commit enables a basic layout for multiple catalogue entries,
the `multiple.html.haml` layout file. A #lookup_entries helper method
has been added to the Catalogue.rb extension file, to handle arrays of
cat numbers.

Also, the cat_details partial has been re-factored to expect a cat
object to be passed in as a parameter. This means the multi-entry layout
can use the same partial, and pass its data to it in a loop.