Use environment vars to generate PDFs
Going forward, derivative formats like PDF and EPUB versions should be
output to the /dist folder. The contents of this folder should not be
added into version control.

This commit enables the generation of a PDF through the use of an
environment flag (-e) at build time. This shoudl be done separately from
deploying the site normally. Copy the derivitave files from /dist and
place them wherever needed.