Add contributor bios
Rather than being written into the page, these are currently stored in
book.yml and then pulled in to the contributors page. The bios may not
be used in other areas of the catalogue, but keeping them in book.yml
is consistent with how we’ve captured and used metadata for the virtual
library, and so may be the best practice we want to follow.