Add correct references formatting (web version)
This commit adds the correct formatting for the References Cited page (2
columns, bibliogrpahic text styles, etc). In order to faciliate this the
underlying markup had to be modified. It's a mix of HTML and Markdown.
One particular note: the line breaks are manually inserted via break
tags after the bibliographic slugs; this seemed like the most explicit
way to represent the content in all areas.

Since the markdown list lives inside of a <div> with the "reference"
class applied, "parse block HTML" has been set to "true" in the
site-wide config.