Switch to svg solution
The original solution in inline styles and css worked at a fixed size
but was not responsive. This svg solution is fully responsive,
maintaining the scale of the object to the comparison person, as well
as the relative scale from entry to entry.

• The scale-mosaic size is linked to the object dimensions in
catalogue.yml as before.
• The scale-person size is based on average human height of 170cm.
• The viewBox width is based on the largest object and in the catalogue
(#2) and would need to be adjusted. This maintains a constant width and
therefore scale from entry to entry while the browser window stays
constant. The height of the viewBox is variable and linked to the
height of the object, or the scale-person in cases were the object
height is smaller.