Add DeepZoom component
This commit enables zoomable images on catalogue entry pages. For now it
only supports single-entry pages; support for multi-entry pages will
come next.

The bulk of this feature lives in the deepzoom.js file. The
functionality follows that of the other leaflet elements as much as
possible (organization and naming of methods, etc). The DeepZoom object
expects to have a cat number and a blob of `zoomData` passed to it upon
initialization. zoomData is an object which should contain the
following properties:
- height: The max height in pixels
- width: max width in pixels
- max_zoom: Max zoom level (will be either 4, 5, or 6)
- min_zoom: Min zoom level (currently this is not being used).

As in the Terracottas publication, the actual image data served here
lives outside of the repository, in a separate Maptiles repo. Use the
following URL schema to access tiles:

http://gettypubs.github.io/mosaics/CAT_NUMBER/{z}/{x}/{y}.png

A plateSetup() function has been added to ui.js to handle the
instantiation of these objects whenever a #plate element is detected.

-----

In addition to implementing this feature, this commit fixes a serious
bug where regional maps were only displaying some of the markers which
should appear. At first I thought the culprit was SmoothState, but it
turns out that there was some kind of collision happening between the
regular map and the PopupMap objects when both appeared on the same
page. In particular, calling the destructive `reverse()` method on the
coordinates object when instantiating a popup map (which was required to
get markers in the right places) seems to have been somehow interfering
with the GeoJSON of the main regional map on the page. The script which
creates popup elements has been modified to keep this from happening;
everything seems to be working normally again.