Use github/master version of middleman-sprockets
See MM issue #1816 for context:
https://github.com/middleman/middleman/issues/1816

For some reason build was failing using the 4.0.0.rc.1 version of
middleman-sprockets on OSX 10.9 (working on 10.10 however). Using the
github master version of this gem allows project to build normally.

Whenever a new stable version of middleman-sprockets is released, the
dependency should be updated to reflect that.