Add print frontmatter page
For various reasons, the best approach for building out the first part
of a print book still involves building out a separate page to come
before all other content. This ensures a really stripped-down HTML
template to feed to Prince. Otherwise there are too many things (and no
easy way to inspect element in PDF) to guarantee a consistent layout.

This commit adds a minimal print-template page and directs the PDF
generation methods to target it first.

Right now this means a /print-template/ page is also available on the
web; this could be removed manually after build, or a line like `rm -rf
/print-template/` could be added into the after_build callback.