Print layout improvements
Additions to the print stylesheet and layout templates produce a better
PDF with Prince. In HTML, some layouts now use a conditional to check if
the app environment is equal to :pdf, and then render accordingly.

A few style changes have also been made to the general (web) page.scss
stylesheet, to handle things like author names in page headers.