Fix print/web layout conflicts
Some of the recent changes on the print and web layouts, respectively,
were interfering with one another. A redundant map partial in the intro
layouts has been removed, and a style update to the print stylesheet (to
reflect the current organization of cat entry pages) fixes a layout
problem in the printed version.