Add partials for About page
These are broken out in such a way as to make them also useful for the
other formats where the About/Copyright page information is often shown
in different order or sometimes not all. The primary standard here has
been to put into partials only that text and information that will be
the same from book to book. Boilerplate. All information specific to
each book is drawn from the YAML so that editors and others working in
the book should only need to look at the yml files in the data
directory, or the markdown files in the source directory, not in the
partials themselves.