Restructure authors mark-up for clarity and citation variants
The previous version of the _cite partial was structured to assume that
citations should always include the primary author along with any noted
contributor, but instead, citations should include only the contributor
as listed, or, the author and contributor, or two contributors
together, if explicitly noted.

This change required restructuring the way we list authors in the page
yaml, which in turn effected some other pages (page layout, contents,
cover, etc) which needed new ways of calling the authors.

For the _cite partial itself, I simplified it significantly using some
variables for the repeating parts. More could be done, but currently
examining the citation as it appears on the About page, and this may
change the solution so held off on further work.