Enable lightbox functionality via PhotoSwipe.js
This commit enables a lightbox feature for the "inline-figure" elements
within a given chapter. By clicking on the image, a responsive lightbox
is triggered that fills the screen with the image, shows the caption,
and allows the user to move between images in gallery mode.

Right now this is pretty much the default functionality provided by
PhotoSwipe.js. This is a pretty sophisticated plugin; it does a lot, so
it makes sense adding this in as a dependency rather than trying to
recreate all of this functionality from scratch.

Some future things to resolve:

- Right now the "X" icon is hidden by the fixed position menu; this
  should be adjusted in the lightbox styles.

- Using left and right arrow keys is supported in this plugin, but these
  keys are already mapped to selecting the next/previous page. There
  needs to be some way to delegate these events so that the keystroke
  does not bubble up when the lightbox is open.