Re-write citation partial
A strange runtime bug in the cite partial was causing the catalogue
entry data on multi-entry layouts to change each time the page was
reloaded - the first entry would disappear, then the second, etc. and
eventually the page would break when no entries remained.

This commit fixes this bug and generally improves the organization of
code when it comes to citations (an area where the combination of
conditional rules and specific strings of text makes things very
tricky). Now all citation related logic has been moved into the
catalogue.rb file as individual helper methods, which are much easier to
read and reason about. The conditional logic remains in the template
sine it determines what should appear on a given page.