Enable figure-specific layout
This commit enables figure-specific layout control on the level of left
or right floating (on web) or full-width, 2 column size (on print).
Classes should be added directly to the figure element in order to
change default display. The classes are:

- .figure-left and .figure-right (for web)
- .figure-full-width (for print)

Also, a .figure-group div can be used to wrap a pair of left/right
figures to ensure they appear normally side by side via clearing rules.
This class is not required, but use where layout problems appear.

An additional change is the addition of new YAML frontmatter on some
pages, a "force_break" property which should be given a value of 'true'
when it exists. This will force the chapter to open on the left-hand
page (assuming a full bleed image on left as the first item inside).
Otherwise chapters flow from page to page normally.