Add popup hover indication
This commit enables an icon on the popup elements which indicates its
type. Not sure if this is going to be jarring from a UI standpoint or
not.

A lot of CSS trickery is required to get this working (this is not a
JS-dependent feature). The basic gist of this is: an ::after
pseudo-element is created for the .popup class. The default state of
this element contains several style rules but nothing that will actually
cause it to display (no width or padding props, for example).

Selected types of popup elements now have a :hover rule which adds
padding to this pseudo-element, along with an appropriate background
image. If more images are created, they should all be 11px x 11px svg
graphics.