Fix flexbox overflow problems in nav sidebar
This commit contains the correct implementation of a bug fix that has
been troubling me for a while. The sidebar nav drawer is set to use
flexbox when at larger screen sizes, because this is the easiest way to
center something vertically.

The problem was, content inside of this sidebar can expand based on user
interaction. While the inner drawers were expanded, the top of the
sidebar got pushed past the extent the user could scroll, chopping off
the book title.

The fix to this takes two parts. First, the entire contents of the
sidebar are now nested under a new div, .nav-sidebar-inner.

This div has a flex rule of: flex: 0 1 auto applied to it. These
settings allow the item to expand when needed without always taking up
all available space. Finally, an overflow: auto rule needs to be applied
to ensure the content is scrollable. A webkit overflow scrolling rule is
necessary for correct behavior on iOS.