Add Points of interest to main maps w/controls
This commit adds a layer control to the region maps, and simple point
markers for all "points of interest" (toggled off by default).

In the future the goal is to allow the user to click on a location link
anywhere in the text and have it:

1. take them to the relevant region map page (if not there already)
2. switch on the points of interest layer
3. read the location out of the URL hash and compare it to the GeoJson
   data to find the desired location
4. center and zoom the map on this location
5. expand the pop-up from this point.

The RegionMap variable has been exposed as a global to allow this from
any page in the site.

Implementation wise, this probably means writing two things:

1. event listeners for hash-change (if user is already on map page) and
   a function in the page-setup SmoothState callback (if user is taken
   to a new page by map link)
2. A function (called by the above) which looks up the hash from the
   GeoJson data and matches it with a feature (probably using _.find()
   or similar), then adds to the map using geometryToLayer(),
   bindPopup(), and openPopup().